Fixed Room.go's, added/implemented Command class
In the hallway, Room.go's were bork'd from a time their classnames were changed (prior to first commit).  References changed to proper class names.

Added `class Command` to engine.py and changed command methods to use it.  By providing a `str` of the player's command to init an instance of this class, it parses the input and divides it into several fields for later use.  Referencing `Command.action` and `Command.target` is easier than parsing for every new command method.  The less string parsing in the long run, the better.